requirements.txt
pandas==1.5.2
requests==2.28.1
flatten-json==0.1.13
oauthlib==3.2.2
python3-linkedin==1.0.2
requests-oauthlib==1.3.1
beautifulsoup4~=4.11.1
PyYAML~=6.0